site/twitterapp/requirements.txt
backports.functools-lru-cache==1.5
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==0.12.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.0
numpy==1.14.2
oauthlib==2.0.6
pyparsing==2.2.0
python-dateutil==2.7.0
python-twitter==3.4.1
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
subprocess32==3.2.7
urllib3==1.22
Werkzeug==0.14.1
gensim==3.4.0
nltk==3.2.5
vaderSentiment==2.5